requirements.txt
Flask==2.3.2
moviepy==1.0.3
Pillow==10.0.0
fpdf2==2.6.0
python-multipart==0.0.6
werkzeug==2.3.7
